extract connection methods to new class to allow reconnect to properly call them
